Should of made sure that worked before pushing it
I hate these yamls some times, litterly missed a single space. Comparsion for bandwith actually works with the numbers being in GB aslong as they are all in GB otherwire it with go to the highest GiB values and pick that. Yes i missed a few gb to gib bout 5 they are  fixed now. No mas for tonight.